I need to figure out how to design the board
The issue I am coming up with now it when rotating the location content on the side columns of the board, They pop out of place.
It's probably the way I designed the board currently but it is an issue.
Also, When you play the game, the player icon changes the look of each part of the board.